fixes #2878 - Import Classes fails when using parser = future
This code is only valid for Puppet >= 3.2
